requirements.txt
corpwechatbot==0.6.2
PyYAML==3.12